visual: rework zoom
Instead of using a set of pre-defined zoom levels that are not
based on the visible area of the window, the zooming action was
very confusing and had some weird behaviors, e.g. when zooming 8x
and the tree actually shrinks.

Fix this misbehavior by adding a set of buttons at the headerbar
and using relative zooming rather than absolute zooming.